#50 Renamed graphics.h to graphics2d.h; removed coord.h and clipping.h; moved the code to graphics2d.h
