+++ graphics/CMakeLists.txt
@@ -21,7 +21,7 @@ link_directories(${GTK3_LIBRARY_DIRS} ${GTKMM3_LIBRARY_DIRS})
 add_definitions(${GTK3_CFLAGS_OTHER} ${GTKMM3_CFLAGS_OTHER})
 
 # Main Target
-set(SOURCE_FILES ui.h graphics.h bezier.h spline.h clipping.h clipping_cs.h clipping_lb.h region.h transforms.h coord.h doubles.h)
+set(SOURCE_FILES ui.h graphics2d.h bezier.h spline.h clipping_cs.h clipping_lb.h region.h transforms.h doubles.h)
 add_executable(graphics main.cpp ${SOURCE_FILES})
 target_link_libraries(graphics ${GTK3_LIBRARIES})
 
@@ -29,4 +29,4 @@ target_link_libraries(graphics ${GTK3_LIBRARIES})
 set(MIN_UNIT_FILES ./tests/min_unit.cpp ./tests/min_unit.h ./tests/timer.cpp ./tests/timer.h)
 add_executable(coord_tests ./tests/coord_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
 add_executable(region_tests tests/region_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
-add_executable(graphics_tests ./tests/graphics_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
+add_executable(graphics2d_tests tests/graphics2d_tests.cpp ${MIN_UNIT_FILES} ${SOURCE_FILES})
