+++ graphics/tests/graphics2d_tests.cpp
@@ -1,5 +1,5 @@
 #include "min_unit.h"
-#include "../graphics.h"
+#include "../graphics2d.h"
 
 static void print(Coord2D coord)
 {
